+++ buildAndReleaseTask/parsescaresults.py
@@ -11,19 +11,38 @@ from junitparser import TestCase, TestSuite, JUnitXml, Failure
 arg_parser = argparse.ArgumentParser(prog="scaresultparser")
 
 # Add the arguments
-arg_parser.add_argument("--target", "-t", metavar="target", type=str,
-                        required=True, help='The target tested')
-arg_parser.add_argument("--mincvss", "-c", metavar="mincvss", type=int,
-                        required=True, default=5,
-                        help='Minimum CVSS score to report on.')
-arg_parser.add_argument("--failbuild", "-f", metavar="failbuild", type=str,
-                        required=True, default="false", choices=["true", "false"],
-                        help='Fail the build (default is false).')
+arg_parser.add_argument(
+    "--target",
+    "-t",
+    metavar="target",
+    type=str,
+    required=True,
+    help="The target tested",
+)
+arg_parser.add_argument(
+    "--mincvss",
+    "-c",
+    metavar="mincvss",
+    type=int,
+    required=True,
+    default=5,
+    help="Minimum CVSS score to report on.",
+)
+arg_parser.add_argument(
+    "--failbuild",
+    "-f",
+    metavar="failbuild",
+    type=str,
+    required=True,
+    default="false",
+    choices=["true", "false"],
+    help="Fail the build (default is false).",
+)
 
 
 def parse_sca_json(data: Dict, min_cvss: int) -> List[Dict]:
-    """ Parse Veracode SCA output
-    JSON schema: https://help.veracode.com/reader/hHHR3gv0wYc2WbCclECf_A/MRCzRYqmVX_PduRZ15UQyA """
+    """Parse Veracode SCA output
+    JSON schema: https://help.veracode.com/reader/hHHR3gv0wYc2WbCclECf_A/MRCzRYqmVX_PduRZ15UQyA"""
 
     results: List[Dict] = []
 
@@ -38,10 +57,10 @@ def parse_sca_json(data: Dict, min_cvss: int) -> List[Dict]:
                 # vulnerability) but two different modules in a project are using it.
                 for library in vuln["libraries"]:
                     result_dict = create_result_dict()
-                    result_dict["Vulnerability"] = vuln['title']
-                    result_dict["CVE"] = vuln['cve']
-                    result_dict["CVSS"] = vuln['cvssScore']
-                    result_dict["Language"] = vuln['language']
+                    result_dict["Vulnerability"] = vuln["title"]
+                    result_dict["CVE"] = vuln["cve"]
+                    result_dict["CVSS"] = vuln["cvssScore"]
+                    result_dict["Language"] = vuln["language"]
                     for key, value in vuln["libraries"][0]["details"][0].items():
                         if key == "updateToVersion":
                             result_dict["Upgrade to Version"] = value
@@ -50,13 +69,23 @@ def parse_sca_json(data: Dict, min_cvss: int) -> List[Dict]:
                         # This gets the ref value
                         library = value.split("/")[4]
                         version = value.split("/")[6]
-                        result_dict["Vulnerable Library"] = data["records"][0]["libraries"][int(library)]["name"]
+                        result_dict["Vulnerable Library"] = data["records"][0][
+                            "libraries"
+                        ][int(library)]["name"]
                         # Version is important because there could be more than one
                         # version installed/used in a project
-                        result_dict["Version"] = data["records"][0]["libraries"][int(library)]["versions"][int(version)]["version"]
+                        result_dict["Version"] = data["records"][0]["libraries"][
+                            int(library)
+                        ]["versions"][int(version)]["version"]
 
                     results.append(result_dict)
                     result_dict = None
+
+        # No vulnerabilities >= the min CVSS score
+        if len(results) == 0:
+            results.append(
+                {"Results": f"No vulnerabilities >= the min CVSS score {min_cvss}."}
+            )
 
     else:
         # No vulnerabilities to address
@@ -66,23 +95,34 @@ def parse_sca_json(data: Dict, min_cvss: int) -> List[Dict]:
 
 
 def create_result_dict() -> Dict:
-    """ Create the results dictionary """
-    return {"Vulnerable Library": None, "Version": None, "Language": None,
-            "Vulnerability": None, "CVE": None, "CVSS": None,
-            "Upgrade to Version": None}
+    """Create the results dictionary"""
+    return {
+        "Vulnerable Library": None,
+        "Version": None,
+        "Language": None,
+        "Vulnerability": None,
+        "CVE": None,
+        "CVSS": None,
+        "Upgrade to Version": None,
+    }
 
 
 def write_output(target: str, results: list) -> None:
-    """ Write scan results in junitxml format """
+    """Write scan results in junitxml format"""
 
     suite = TestSuite(f"{target}")
 
     for result in results:
         if result != {"Results": "No vulnerabilities."}:
             test_case = TestCase(result["Vulnerable Library"])
-            test_case.name = (result["Vulnerable Library"] + " - "\
-                 + result["Vulnerability"] + " - "\
-                 + "CVSS " + str(result["CVSS"]))
+            test_case.name = (
+                result["Vulnerable Library"]
+                + " - "
+                + result["Vulnerability"]
+                + " - "
+                + "CVSS "
+                + str(result["CVSS"])
+            )
             test_case.result = [Failure(result)]
         else:
             test_case = TestCase("No vulnerabilities")
@@ -92,21 +132,22 @@ def write_output(target: str, results: list) -> None:
 
     xml = JUnitXml()
     xml.add_testsuite(suite)
-    xml.write('test-output.xml')
+    xml.write("test-output.xml")
 
 
 def get_parent_package(transitive_package: str) -> str:
-    """ npm list command: submit the vulnerable package
+    """npm list command: submit the vulnerable package
     to get the top level package that needs to be patched.
-    I still need to do npm ci --ignore-scripts before doing this """
+    I still need to do npm ci --ignore-scripts before doing this"""
 
     list_command = f"npm list {transitive_package} --json"
 
     _check = None
 
     try:
-        _check = subprocess.run(list_command, capture_output=True,
-                                shell=True, check=True)
+        _check = subprocess.run(
+            list_command, capture_output=True, shell=True, check=True
+        )
 
     except subprocess.SubprocessError as err:
         print(err)  # This will fail if it can't find npm on the system
@@ -116,15 +157,15 @@ def get_parent_package(transitive_package: str) -> str:
 
     # Get the top-level package
     top_package = ""
-    for key, value in npm_output['dependencies'].items():
-        top_package = (value['from'])
+    for key, value in npm_output["dependencies"].items():
+        top_package = value["from"]
 
     return top_package
 
 
 def main() -> None:
 
-    """ Main function """
+    """Main function"""
     # Execute the parse_args() method
     args = arg_parser.parse_args()
     target = args.target
@@ -132,7 +173,7 @@ def main() -> None:
     fail_build = args.failbuild
 
     # Open the Veracode SCA JSON results
-    with open('scaresults.json', 'r') as sca_results:
+    with open("scaresults.json", "r") as sca_results:
         data = json.load(sca_results)
     # Parse results
     results = parse_sca_json(data, min_cvss)
@@ -141,9 +182,11 @@ def main() -> None:
     write_output(target, results)
 
     # Remove scaresults.json file
-    os.remove('scaresults.json')
+    os.remove("scaresults.json")
 
-    output = os.path.normpath(os.path.abspath(os.path.expanduser(os.path.expandvars("test-output.xml"))))
+    output = os.path.normpath(
+        os.path.abspath(os.path.expanduser(os.path.expandvars("test-output.xml")))
+    )
 
     # Borrowed from pytest-azurepipelines
     # https://github.com/tonybaloney/pytest-azurepipelines/blob/master/pytest_azurepipelines.py